requirements.txt
aiohttp==3.9.1
google-auth==2.28.1
google-cloud-bigquery==3.20.1
pandas==2.1.4
numpy==1.26.3